Move raiden.utils.filter to raiden.blockchain.filter
