+++ raiden/network/proxies/payment_channel.py
@@ -2,11 +2,11 @@ from typing import Optional
 
 from web3.utils.filters import Filter
 
+from raiden.blockchain.filters import decode_event, get_filter_args_for_specific_event_from_channel
 from raiden.constants import UINT256_MAX
 from raiden.network.proxies.token_network import ChannelDetails, TokenNetwork
 from raiden.network.proxies.utils import get_channel_participants_from_open_event
 from raiden.transfer.state import PendingLocksState
-from raiden.utils.filters import decode_event, get_filter_args_for_specific_event_from_channel
 from raiden.utils.typing import (
     AdditionalHash,
     Address,
